Reset structure of project and update requirements and .gitignore
