+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__/
 /venv/
 .tox/
 *.egg-info/
+db.sqlite3
