+++ requirements-dev.txt
@@ -1,3 +1,4 @@
 pytest
 flask
 pyflakes
+Django
