+++ tests/fixtures/demo_project_django/demo_app/urls.py
@@ -5,4 +5,5 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
+    path('second_page.html', views.second_page, name='second_page'),
 ]
