--- tests/fixtures/demo_project_django/demo_app_2pages/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class DemoApp2PagesConfig(AppConfig):
-    name = 'demo_app_2pages'
