--- tests/fixtures/demo_project_django/demo_app_2pages/tests.py
@@ -1,3 +1,0 @@
-from django.test import TestCase
-
-# Create your tests here.
