--- tests/fixtures/demo_project_django/demo_app_2pages/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from . import views
-
-
-urlpatterns = [
-    path('', views.index, name='index'),
-    path('second_page.html', views.second_page, name='second_page'),
-]
